Implemented consumeKey and group trust
Also, cleaned up code a bit and ran analysis, taking some of it's
suggestions.
Could do with a bit of refactoring now, as it's become a bit of a
spaghetti mess, but I can't be arsed as long as it works perfectly and
isn't slow.
Will revisit refactoring if it becomes hard to implement stuff, or bugs
crop up.