Got rid of LeftLock, added ProtectiveAuras, etc
- LeftLock
Now that there's no more left click interactions on doors etc, it
makes very little sense to have that configurable.

- Protective auras
Disabled by default, 'useProtectiveAuras: true' in the configuration
will now check for locked blocks around placed blocks and stop the
placement if that locked block isn't yours, but is listed in the
protectiveAura section of the configuration.
By default, this is CHEST and DISPENSER.
The point of this feature is to stop tekkit-tubes from sucking items
from locked chests.
It can be a little CPU intensive as the check happens on every single
block place, so if you get CPU spikes, turn this off.

- etc
Added BEACON to default lockables list.
Tiiiny little code changes, mostly just indenting and stuff